requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
bleach==6.2.0
celery==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
crispy-bootstrap5==0.7
cryptography==44.0.2
defusedxml==0.7.1
Django==4.1.13
django-allauth==0.52.0
django-bootstrap5==22.1
django-ckeditor-5==0.2.17
django-crispy-forms==2.0
django-debug-toolbar==4.3.0
django-environ==0.12.0
django-js-asset==2.2.0
django-quill-editor==0.1.42
django-ratelimit==4.1.0
django-tables==0.2
et_xmlfile==2.0.0
flower==2.0.1
gevent==25.4.1
greenlet==3.2.0
humanize==4.12.2
idna==3.10
kombu==5.5.2
numpy==2.0.2
oauthlib==3.2.2
openpyxl==3.1.5
pandas==2.2.3
pillow==11.2.1
prometheus_client==0.21.1
prompt_toolkit==3.0.50
psycopg-binary==3.2.6
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python3-openid==3.2.0
pytz==2025.2
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
sqlparse==0.5.3
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.4.0
uuid==1.30
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
zope.event==5.0
zope.interface==7.2
